refactor: move electron browser view event listeners into separate module
